requirements.txt
python-telegram-bot
sentence-transformers
numpy
ollama
Pillow
transformers
torch
